bug:agent和引擎并发执行task任务时可能造成task状态出错 #8465
