+++ src/backend/ci/core/process/biz-engine/src/main/kotlin/com/tencent/devops/process/engine/control/TaskControl.kt
@@ -32,7 +32,6 @@ import com.tencent.devops.common.api.util.Watcher
 import com.tencent.devops.common.api.util.timestampmilli
 import com.tencent.devops.common.event.dispatcher.pipeline.PipelineEventDispatcher
 import com.tencent.devops.common.event.enums.ActionType
-import com.tencent.devops.common.log.utils.BuildLogPrinter
 import com.tencent.devops.common.pipeline.enums.BuildStatus
 import com.tencent.devops.common.pipeline.pojo.element.RunCondition
 import com.tencent.devops.common.redis.RedisOperation
@@ -42,19 +41,15 @@ import com.tencent.devops.process.engine.atom.AtomResponse
 import com.tencent.devops.process.engine.atom.TaskAtomService
 import com.tencent.devops.process.engine.common.BS_ATOM_STATUS_REFRESH_DELAY_MILLS
 import com.tencent.devops.process.engine.common.BS_TASK_HOST
-import com.tencent.devops.process.engine.common.Timeout
 import com.tencent.devops.process.engine.control.lock.ContainerIdLock
 import com.tencent.devops.process.engine.pojo.PipelineBuildTask
 import com.tencent.devops.process.engine.pojo.event.PipelineBuildAtomTaskEvent
-import com.tencent.devops.process.engine.pojo.event.PipelineBuildContainerEvent
+import com.tencent.devops.process.engine.service.PipelineBuildTaskService
 import com.tencent.devops.process.engine.service.PipelineRuntimeService
 import com.tencent.devops.process.engine.service.PipelineTaskService
-import com.tencent.devops.process.engine.utils.BuildUtils
-import com.tencent.devops.process.util.TaskUtils
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import java.util.concurrent.TimeUnit
 
 /**
  * 任务（最小单元Atom）控制器
@@ -67,7 +62,7 @@ class TaskControl @Autowired constructor(
     private val pipelineEventDispatcher: PipelineEventDispatcher,
     private val pipelineRuntimeService: PipelineRuntimeService,
     private val pipelineTaskService: PipelineTaskService,
-    private val buildLogPrinter: BuildLogPrinter
+    private val pipelineBuildTaskService: PipelineBuildTaskService
 ) {
 
     companion object {
@@ -153,7 +148,12 @@ class TaskControl @Autowired constructor(
                     startTime = buildTask.startTime?.timestampmilli() ?: System.currentTimeMillis(),
                     atomResponse = atomResponse
                 )
-                return finishTask(buildTask, buildStatus)
+                pipelineBuildTaskService.finishTask(
+                    buildTask = buildTask,
+                    buildStatus = buildStatus,
+                    actionType = actionType,
+                    source = source
+                )
             }
         } else {
             buildTask.starter = userId
@@ -170,7 +170,12 @@ class TaskControl @Autowired constructor(
                 // 如果是要轮循的才需要定时消息轮循
                 loopDispatch(buildTask = buildTask)
             } else {
-                finishTask(buildTask = buildTask, buildStatus = buildStatus)
+                pipelineBuildTaskService.finishTask(
+                    buildTask = buildTask,
+                    buildStatus = buildStatus,
+                    actionType = actionType,
+                    source = source
+                )
             }
         }
     }
@@ -222,132 +227,7 @@ class TaskControl @Autowired constructor(
         pipelineEventDispatcher.dispatch(this)
     }
 
-    /**
-     * 对结束的任务[buildTask], 根据状态[buildStatus]是否失败，以及[buildTask]配置：
-     * 1. 需要失败重试，将[buildTask]的构建状态设置为RETRY
-     */
-    private fun PipelineBuildAtomTaskEvent.finishTask(buildTask: PipelineBuildTask, buildStatus: BuildStatus) {
-        if (buildStatus == BuildStatus.CANCELED) {
-            // 删除redis中取消构建操作标识
-            redisOperation.delete(BuildUtils.getCancelActionBuildKey(buildId))
-            redisOperation.delete(TaskUtils.getCancelTaskIdRedisKey(buildId, containerId, false))
-            // 当task任务是取消状态时，把taskId存入redis供心跳接口获取
-            val cancelTaskKey = TaskUtils.getCancelTaskIdRedisKey(buildId, containerId)
-            redisOperation.leftPush(cancelTaskKey, taskId)
-            // 为取消任务设置最大超时时间，防止构建异常产生的脏数据
-            redisOperation.expire(cancelTaskKey, TimeUnit.DAYS.toSeconds(Timeout.MAX_JOB_RUN_DAYS))
-        }
-        // 如果是取消的构建，则会统一取消子流水线的构建
-        if (buildStatus.isPassiveStop() || buildStatus.isCancel()) {
-            terminateSubPipeline(buildId, buildTask)
-        }
-        // 失败的任务并且不是需要前置终止的情况才允许自动重试
-        if (buildStatus.isFailure() && !actionType.isTerminate() && !FastKillUtils.isTerminateCode(errorCode)) {
-            // 如果配置了失败重试，且重试次数上线未达上限，则将状态设置为重试，让其进入
-            if (pipelineTaskService.isRetryWhenFail(buildTask.projectId, taskId, buildId)) {
-                LOG.info("ENGINE|$buildId|$source|ATOM_FIN|$stageId|j($containerId)|t($taskId)|RetryFail")
-                pipelineTaskService.updateTaskStatus(
-                    task = buildTask, userId = buildTask.starter, buildStatus = BuildStatus.RETRY
-                )
-            } else {
-                // 如果配置了失败继续，则继续下去的行为是在ContainerControl处理，而非在Task
-                pipelineTaskService.createFailTaskVar(
-                    buildId = buildId,
-                    projectId = projectId,
-                    pipelineId = pipelineId,
-                    taskId = taskId
-                )
-            }
-        } else {
-            // 清除该原子内的重试记录
-            pipelineTaskService.removeRetryCache(buildId, taskId)
-            // 清理插件错误信息（重试插件成功的情况下）
-            pipelineTaskService.removeFailTaskVar(
-                buildId = buildId,
-                projectId = projectId,
-                pipelineId = pipelineId,
-                taskId = taskId
-            )
-        }
-
-        if (errorTypeName != null && ErrorType.getErrorType(errorTypeName!!) != null) {
-            pipelineTaskService.setTaskErrorInfo(
-                transactionContext = null,
-                projectId = projectId,
-                buildId = buildId,
-                taskId = taskId,
-                errorCode = errorCode,
-                errorType = ErrorType.getErrorType(errorTypeName!!)!!,
-                errorMsg = reason ?: "unknown"
-            )
-        }
-
-        pipelineEventDispatcher.dispatch(
-            PipelineBuildContainerEvent(
-                source = "from_t($taskId)",
-                projectId = projectId,
-                pipelineId = pipelineId,
-                userId = userId,
-                buildId = buildId,
-                stageId = stageId,
-                containerId = containerId,
-                containerHashId = containerHashId,
-                containerType = containerType,
-                actionType = actionType,
-                errorCode = errorCode,
-                errorTypeName = errorTypeName,
-                reason = reason
-            )
-        )
-    }
-
     private fun atomBuildStatus(response: AtomResponse): BuildStatus {
         return response.buildStatus
-    }
-
-    private fun terminateSubPipeline(buildId: String, buildTask: PipelineBuildTask) {
-
-        if (buildTask.subBuildId.isNullOrBlank()) {
-            return
-        }
-
-        val subBuildInfo = pipelineRuntimeService.getBuildInfo(buildTask.subProjectId!!, buildTask.subBuildId!!)
-
-        if (subBuildInfo?.status?.isFinish() == false) { // 子流水线状态为未构建结束的，开始下发退出命令
-            try {
-                val tasks = pipelineTaskService.getRunningTask(subBuildInfo.projectId, subBuildInfo.buildId)
-                tasks.forEach { task ->
-                    val taskId = task["taskId"] ?: ""
-                    val containerId = task["containerId"] ?: ""
-                    val executeCount = task["executeCount"] ?: 1
-                    buildLogPrinter.addYellowLine(
-                        buildId = buildId,
-                        message = "Cancelled by pipeline[${buildTask.pipelineId}]，Operator:${buildTask.starter}",
-                        tag = taskId.toString(),
-                        jobId = containerId.toString(),
-                        executeCount = executeCount as Int
-                    )
-                }
-
-                if (tasks.isEmpty()) {
-                    buildLogPrinter.addYellowLine(
-                        buildId = buildId,
-                        message = "cancelled by pipeline[${buildTask.pipelineId}]，Operator:${buildTask.starter}",
-                        tag = "",
-                        jobId = "",
-                        executeCount = 1
-                    )
-                }
-                pipelineRuntimeService.cancelBuild(
-                    projectId = subBuildInfo.projectId,
-                    pipelineId = subBuildInfo.pipelineId,
-                    buildId = subBuildInfo.buildId,
-                    userId = subBuildInfo.startUser,
-                    buildStatus = BuildStatus.CANCELED
-                )
-            } catch (ignored: Exception) {
-                LOG.warn("ENGINE|$buildId|TerminateSubPipeline|subBuildId=${subBuildInfo.buildId}|e=$ignored")
-            }
-        }
     }
 }
